src/models/schemas.py
from pydantic import BaseModel
from typing import Optional, List

class ChatRequest(BaseModel):
    message: str
    user_id: str = "anonymous"

class ChatResponse(BaseModel):
    reply: str
    mood: Optional[str] = None
    recommendations: Optional[List[str]] = None
