requirements.txt
python-dotenv==1.2.1
langchain==1.0.5
langchain-groq==1.0.0
fastapi==0.121.0
requests==2.32.5
spotipy==2.25.1
redis==7.0.1
uvicorn==0.38.0
pytest==9.0.0